requirements.txt
langchain 
wikipedia 
langchain-community
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
ddgs
reportlab